first iteration, reg screen widgets + one implemented screen
