+++ aux_ui/lib/main.dart
@@ -1,10 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:aux_ui/theme/aux_theme.dart';
-import './widgets/aux_card.dart';
-
+import 'package:aux_ui/screens/guest_reg.dart';
 void main() => runApp(AuxApp());
-
-
 
 class AuxApp extends StatelessWidget {
   // This widget is the root of your application.
@@ -19,111 +16,7 @@ class AuxApp extends StatelessWidget {
         fontFamily: 'Larsseit',
         textTheme: auxTextTheme, // TODO add secondary text theme for alt buttons?
       ),
-      home: GuestRegName(title: 'aux'),
-    );
-  }
-}
-
-class GuestRegName extends StatefulWidget {
-  GuestRegName({Key key, this.title}) : super(key: key);
-  final String title;
-  @override
-  _GuestRegState createState() => _GuestRegState();
-}
-
-class _GuestRegState extends State<GuestRegName> {
-  @override
-  Widget build(BuildContext context) {
-    SizeConfig().init(context);
-    return Container(
-      height: SizeConfig.safeBlockVertical * 100,
-      width: SizeConfig.safeBlockHorizontal * 100,
-      color: auxPrimary,
-      child: Padding(
-          padding: SizeConfig.notchPadding,
-          child: AuxCard(
-            borderColor: auxAccent,
-            padding: 24.0,
-            child: Column(
-              mainAxisSize: MainAxisSize.max,
-              children: <Widget>[
-                Expanded(
-                  child: Padding(
-                    padding: EdgeInsets.only(bottom: 20),
-                    child: Align(
-                      alignment: Alignment.bottomLeft,
-                      child: Text(
-                        "first,\nlet's get a name!", 
-                        style: auxDisp3
-                      )
-                    )
-                  )
-                ),
-                Expanded(
-                  child: Padding(
-                    padding: EdgeInsets.only(top: 20),
-                    child: Align(
-                      alignment: Alignment.topCenter,
-                      child: Column(
-                        children: <Widget>[
-                          TextField(
-                            style: auxAccentButton,
-                            textAlignVertical: TextAlignVertical.bottom,
-                            decoration: InputDecoration(
-                              prefixIcon: Icon(
-                                Icons.short_text, 
-                                color:auxAccent,
-                                size: 26.0,
-                                semanticLabel: "Short text for user nickname",
-                              ),
-                              labelStyle: auxAccentButton,
-                              labelText: 'enter a nickname',
-                              enabledBorder: OutlineInputBorder(
-                                borderSide: BorderSide(color: Colors.white, width: 3),
-                                borderRadius: BorderRadius.circular(10), // TODO: scale by screen resolution
-                              ),
-                              focusedBorder: OutlineInputBorder(
-                                borderSide: BorderSide(color: Colors.white, width: 3),
-                                borderRadius: BorderRadius.circular(10), // TODO: scale by screen resolution
-                              ),
-                            ),
-                          ),
-                          Padding(
-                            padding: EdgeInsets.all(20),
-                            child: Text('or', style: auxAccentButton)
-                          ),
-                          ButtonTheme(
-                            minWidth: SizeConfig.safeBlockHorizontal * 100,
-                            child: FlatButton(
-                              onPressed:() {},
-                              color: auxAccent,
-                              padding: EdgeInsets.all(15),
-                              shape: RoundedRectangleBorder(
-                                borderRadius: BorderRadius.circular(10)
-                              ),
-                              child: Stack(
-                                children: <Widget>[
-                                  Align(
-                                    alignment: Alignment.centerLeft,
-                                    child: Image.asset('assets/spotify_logo.png', height: 21, width: 21),
-                                  ),
-                                  Align(
-                                    alignment: Alignment.center,
-                                    child: Text('sign up with spotify', strutStyle: StrutStyle(fontFamily: 'Larsseit', fontSize: 17, height: 1.2) ),
-                                  ),
-                                ],
-                              ) 
-                            ),
-                          )
-                        ],
-                      )
-                    )
-                  )
-                )
-              ], 
-            ),
-          )
-        )
+      home: GuestReg(),
     );
   }
 }
